added graph function to own class in own file
